Fix Cppcheck 2.22 diagnostics
* Fix Cppcheck 2.22 diagnostics

Cppcheck 2.22 began reporting failures in native code that 2.21
accepted, including actionable warnings and parser errors in invalid
preprocessor configurations. The temporary 2.21 pin left CI behind
the newer analyzer.

Pin 2.22 in GitLab and GitHub CI, correct the native warnings, and
exclude configurations that Cppcheck cannot parse. Keep checks active
for supported build paths and restore the CI gate without changing
runtime behavior.


Approved-by: Eric Shi <ershi@nvidia.com>

See merge request omniverse/warp!3052

+++ warp/native/mesh.h
@@ -2494,7 +2494,10 @@ CUDA_CALLABLE inline vec3 mesh_get_point(uint64_t id, int index)
 
 #if FP_CHECK
     if (index >= mesh.num_tris * 3) {
-        printf("mesh_get_point (%llu, %d) out of bounds at %s:%d\n", id, index, __FILE__, __LINE__);
+        printf(
+            "mesh_get_point (%llu, %d) out of bounds at %s:%d\n", static_cast<unsigned long long>(id), index, __FILE__,
+            __LINE__
+        );
         assert(0);
     }
 #endif
@@ -2519,7 +2522,10 @@ CUDA_CALLABLE inline vec3 mesh_get_velocity(uint64_t id, int index)
 
 #if FP_CHECK
     if (index >= mesh.num_tris * 3) {
-        printf("mesh_get_velocity (%llu, %d) out of bounds at %s:%d\n", id, index, __FILE__, __LINE__);
+        printf(
+            "mesh_get_velocity (%llu, %d) out of bounds at %s:%d\n", static_cast<unsigned long long>(id), index,
+            __FILE__, __LINE__
+        );
         assert(0);
     }
 #endif
